show person name in newly added organisation membership
If we are adding an organization membership then the newly added listing
should show the name of the person. To do this pass in the object type
to the membership dialog and then use that to populate the correct key
and also pass it back to the item template so it does the right thing.

Fixes #467